requirements.txt
action-msgs==1.2.1
action-tutorials-interfaces==0.20.4
action-tutorials-py==0.20.4
actionlib-msgs==4.2.4
altair==5.3.0
ament-cmake-test==1.3.9
ament-copyright==0.12.11
ament-cppcheck==0.12.11
ament-cpplint==0.12.11
ament-flake8==0.12.11
ament-index-python==1.4.0
ament-lint==0.12.11
ament-lint-cmake==0.12.11
ament-package==0.14.0
ament-pep257==0.12.11
ament-uncrustify==0.12.11
ament-xmllint==0.12.11
angles==1.15.0
annotated-types==0.7.0
attrs==23.2.0
blinker==1.8.2
builtin-interfaces==1.2.1
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
composition-interfaces==1.2.1
control-msgs==4.6.0
controller-manager==2.41.0
controller-manager-msgs==2.41.0
cv-bridge==3.2.1
demo-nodes-py==0.20.4
diagnostic-msgs==4.2.4
domain-coordinator==0.10.0
easyocr==1.7.1
example-interfaces==0.9.3
examples-rclpy-executors==0.15.1
examples-rclpy-minimal-action-client==0.15.1
examples-rclpy-minimal-action-server==0.15.1
examples-rclpy-minimal-client==0.15.1
examples-rclpy-minimal-publisher==0.15.1
examples-rclpy-minimal-service==0.15.1
examples-rclpy-minimal-subscriber==0.15.1
filelock==3.14.0
fsspec==2024.5.0
gazebo-msgs==3.7.0
generate-parameter-library-py==0.3.8
geometry-msgs==4.2.4
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.131.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.6.0
googleapis-common-protos==1.63.1
grpcio==1.64.0
grpcio-status==1.62.2
httplib2==0.22.0
idna==3.7
image-geometry==3.2.1
imageio==2.34.1
interactive-markers==2.3.2
Jinja2==3.1.4
joint-state-publisher==2.4.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
laser-geometry==2.4.0
launch==1.0.6
launch-param-builder==0.1.1
launch-ros==0.19.7
launch-testing==1.0.6
launch-testing-ros==0.19.7
launch-xml==1.0.6
launch-yaml==1.0.6
lazy_loader==0.4
lifecycle-msgs==1.2.1
logging-demo==0.20.4
map-msgs==2.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
message-filters==4.3.4
moveit-configs-utils==2.5.5
moveit-msgs==2.2.1
mpmath==1.3.0
nav-msgs==4.2.4
networkx==3.3
ninja==1.11.1.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
object-recognition-msgs==2.0.0
octomap-msgs==2.0.0
opencv-python-headless==4.10.0.82
osrf-pycommon==2.0.2
oss-interfaces==5.3.2
packaging==24.0
pandas==2.2.2
pcl-msgs==1.0.0
pdf2image==1.17.0
pendulum-msgs==0.20.4
pillow==10.3.0
poppler-utils==0.1.0
proto-plus==1.23.0
protobuf==4.25.3
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyclipper==1.3.0.post5
pydantic==2.7.2
pydantic_core==2.18.3
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.2
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-qt-binding==1.1.2
pytz==2024.1
PyYAML==6.0.1
qt-dotgraph==2.2.3
qt-gui==2.2.3
qt-gui-cpp==2.2.3
qt-gui-py-common==2.2.3
quality-of-service-demo-py==0.20.4
rcl-interfaces==1.2.1
rclpy==3.3.13
rcutils==5.1.6
referencing==0.35.1
requests==2.32.3
resource-retriever==3.1.2
rich==13.7.1
rmw-dds-common==1.6.0
ros2action==0.18.10
ros2bag==0.15.11
ros2cli==0.18.10
ros2component==0.18.10
ros2doctor==0.18.10
ros2interface==0.18.10
ros2launch==0.19.7
ros2lifecycle==0.18.10
ros2multicast==0.18.10
ros2node==0.18.10
ros2param==0.18.10
ros2pkg==0.18.10
ros2run==0.18.10
ros2service==0.18.10
ros2topic==0.18.10
rosbag2-interfaces==0.15.11
rosbag2-py==0.15.11
rosgraph-msgs==1.2.1
rosidl-adapter==3.1.5
rosidl-cli==3.1.5
rosidl-cmake==3.1.5
rosidl-generator-c==3.1.5
rosidl-generator-cpp==3.1.5
rosidl-generator-py==0.14.4
rosidl-parser==3.1.5
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.1
rosidl-typesupport-cpp==2.0.1
rosidl-typesupport-fastrtps-c==2.2.2
rosidl-typesupport-fastrtps-cpp==2.2.2
rosidl-typesupport-introspection-c==3.1.5
rosidl-typesupport-introspection-cpp==3.1.5
rpds-py==0.18.1
rpyutils==0.2.1
rqt-action==2.0.1
rqt-bag==1.1.5
rqt-bag-plugins==1.1.5
rqt-console==2.0.3
rqt-graph==1.3.1
rqt-gui==1.1.7
rqt-gui-py==1.1.7
rqt-msg==1.2.0
rqt-plot==1.1.2
rqt-publisher==1.5.0
rqt-py-common==1.1.7
rqt-py-console==1.0.2
rqt-reconfigure==1.1.2
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-topic==1.5.0
rsa==4.9
scikit-image==0.23.2
scipy==1.13.1
scripts==3.7.0
sensor-msgs==4.2.4
sensor-msgs-py==4.2.4
shape-msgs==4.2.4
shapely==2.0.4
six==1.16.0
smmap==5.0.1
srdfdom==2.0.4
sros2==0.10.5
statistics-msgs==1.2.1
std-msgs==4.2.4
std-srvs==4.2.4
stereo-msgs==4.2.4
streamlit==1.35.0
sympy==1.12.1
teleop-twist-keyboard==2.4.0
tenacity==8.3.0
tf2-geometry-msgs==0.25.7
tf2-kdl==0.25.7
tf2-msgs==0.25.7
tf2-py==0.25.7
tf2-ros-py==0.25.7
tf2-tools==0.25.7
tifffile==2024.5.22
toml==0.10.2
toolz==0.12.1
topic-monitor==0.20.4
torch==2.3.0
torchvision==0.18.0
tornado==6.4
tqdm==4.66.4
trajectory-msgs==4.2.4
triton==2.3.0
turtlesim==1.4.2
typing_extensions==4.12.1
tzdata==2024.1
# Editable install with no version control (uf-ros-lib==0.0.0)
-e /home/sanchay/ros2_workspaces/oss_ws/build/uf_ros_lib
unique-identifier-msgs==2.2.1
urdfdom-py==1.2.1
uritemplate==4.1.1
urllib3==2.2.1
visualization-msgs==4.2.4
watchdog==4.0.1
xacro==2.0.8
xarm-msgs==0.0.0
